# testdata/openapi3_security/docs/swagger.json
{
    "components": {
        "securitySchemes": {
            "ApiKeyAuth": {
                "description": "API key authentication",
                "in": "header",
                "name": "X-API-Key",
                "type": "apiKey"
            },
            "BasicAuth": {
                "description": "HTTP Basic Authentication",
                "scheme": "basic",
                "type": "http"
            },
            "BearerAuth": {
                "bearerFormat": "JWT",
                "description": "JWT Bearer token authentication",
                "scheme": "bearer",
                "type": "http"
            },
            "OAuth2": {
                "flows": {
                    "authorizationCode": {
                        "authorizationUrl": "https://example.com/oauth/authorize",
                        "scopes": {
                            "admin": "Grants read and write access to administrative information",
                            "read": "Grants read access",
                            "write": "Grants write access"
                        },
                        "tokenUrl": "https://example.com/oauth/token"
                    }
                },
                "type": "oauth2"
            },
            "OpenIDConnect": {
                "description": "OpenID Connect Discovery",
                "openIdConnectUrl": "https://example.com/.well-known/openid-configuration",
                "type": "openIdConnect"
            }
        }
    },
    "info": {
        "contact": {
            "email": "support@swagger.io",
            "name": "API Support",
            "url": "http://www.swagger.io/support"
        },
        "description": "This is a sample server demonstrating OpenAPI 3.0 security schemes.",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "termsOfService": "http://swagger.io/terms/",
        "title": "Swagger Example API with OpenAPI 3.0 Security",
        "version": "1.0"
    },
    "openapi": "3.0.0",
    "paths": {
        "/apikey": {
            "get": {
                "description": "This endpoint requires API key in header",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "ok"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "unauthorized"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "summary": "API key endpoint",
                "tags": [
                    "auth"
                ]
            }
        },
        "/basic": {
            "get": {
                "description": "This endpoint requires HTTP Basic authentication",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "ok"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "unauthorized"
                    }
                },
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "summary": "Basic auth endpoint",
                "tags": [
                    "auth"
                ]
            }
        },
        "/bearer": {
            "get": {
                "description": "This endpoint requires JWT Bearer token",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "ok"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "unauthorized"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Bearer token endpoint",
                "tags": [
                    "auth"
                ]
            }
        },
        "/oauth": {
            "get": {
                "description": "This endpoint requires OAuth2 authentication",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "ok"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "unauthorized"
                    }
                },
                "security": [
                    {
                        "OAuth2": [
                            "read",
                            "write"
                        ]
                    }
                ],
                "summary": "OAuth2 endpoint",
                "tags": [
                    "auth"
                ]
            }
        },
        "/openid": {
            "get": {
                "description": "This endpoint requires OpenID Connect authentication",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "ok"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "unauthorized"
                    }
                },
                "security": [
                    {
                        "OpenIDConnect": []
                    }
                ],
                "summary": "OpenID Connect endpoint",
                "tags": [
                    "auth"
                ]
            }
        },
        "/public": {
            "get": {
                "description": "This endpoint does not require authentication",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "ok"
                    }
                },
                "summary": "Public endpoint",
                "tags": [
                    "public"
                ]
            }
        }
    },
    "servers": [
        {
            "url": "https://localhost:8080/v1"
        }
    ]
}